项目

+++ finance/subject/views.py
@@ -1,9 +1,12 @@
+import json
 from decimal import Decimal
 from datetime import date, datetime
 from rest_framework.views import APIView
 from .models import *
+from .func import get_all_subject
 from user.models import User
-from user.func import is_logined
+from user.func import is_logined, has_permission
+from subject.func import get_all_subject
 from finance.basic import common_response
 from utils.tools import HASHIDS
 
@@ -19,6 +22,10 @@ class SetOfAccountsView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'SetOfAccounts', 'POST'):
+            return common_response(code=500, msg='您没有操作权限')
+
         data = request.data
         name = data.get('name', '')
         soa_date = data.get('date', date.today())
@@ -32,6 +39,14 @@ class SetOfAccountsView(APIView):
             date=soa_date,
             village=obj.village
         )
+        # 初始化科目余额
+        for sub_obj in Subject.objects.filter(village=None):
+            subjectofaccounts_obj = SubjectOfAccounts.objects.create(
+                account=soa_obj,
+                subject=sub_obj
+            )
+            subjectofaccounts_obj.save()
+
         soa_obj.save()
         return common_response(msg='True')
 
@@ -43,9 +58,9 @@ class SetOfAccountsView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
-        # 只有管理员可修改
-        if obj.identity != 0:
-            return common_response(code=500, msg='当前身份无权修改')
+        # 验证用户权限
+        if not has_permission(obj, 'SetOfAccounts', 'PUT'):
+            return common_response(code=500, msg='您没有操作权限')
 
         data = request.data
         soa_id = data.get('soa_id', '')
@@ -81,6 +96,10 @@ class SetOfAccountsView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(user_obj, 'SetOfAccounts', 'GET'):
+            return common_response(code=500, msg='您没有操作权限')
+
         soa_id = request.GET.get('soa_id', '')
         if soa_id:
             try:
@@ -95,7 +114,7 @@ class SetOfAccountsView(APIView):
             }
             return common_response(data=soa_info)
 
-        get_type = request.GET.get('type', '')
+        get_type = request.GET.get('type', 'list')
         if get_type == 'list':
             soa_list = []
             for obj in SetOfAccounts.objects.all():
@@ -121,12 +140,16 @@ class SubjectView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Subject', 'POST'):
+            return common_response(code=500, msg='您没有操作权限')
+
         data = request.data
         # 验证数据
         subject_num = data.get('subject_num', '')
         subject_name = data.get('subject_name', '')
         subject_father = data.get('subject_father', None)
-        subject_type =data.get('subject_type', 0)
+        subject_type = data.get('subject_type', 0)
         balance_type = data.get('balance_type', 0)
         assists = data.get('assists', 0)
         soa_id = data.get('soa_id', 0)
@@ -151,7 +174,9 @@ class SubjectView(APIView):
             subject_name=subject_name,
             parent=subject_father,
             subject_type=subject_type,
-            balance_type=balance_type
+            balance_type=balance_type,
+            can_operation=True,
+            village=obj.village
         )
         # 存储多对多关联表数据
         if assists:
@@ -161,13 +186,12 @@ class SubjectView(APIView):
             subject_obj.assist_business.add(*ass_obj)
 
         # 科目余额
-        num = data.get('num', 0)
-        money = Decimal(data.get('money', 0)).quantize(Decimal('0.00'))
         soa_obj = SubjectOfAccounts.objects.create(
             account_id=soa_id,
             subject=subject_obj,
-            num=num,
-            money=money
+            num=0,
+            money=0,
+            balance=0
         )
 
         subject_obj.save()
@@ -182,12 +206,20 @@ class SubjectView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Subject', 'DELETE'):
+            return common_response(code=500, msg='您没有操作权限')
+
         data = request.data
         sub_id = data.get('sub_id', 0)
         try:
             sub_obj = Subject.objects.get(id=sub_id)
         except Subject.DoesNotExist:
             return common_response(code=500, msg='ID不存在')
+
+        # 判断当前用户是否可操作
+        if not sub_obj.can_operation:
+            return common_response(code=500, msg='当前科目不可操作')
 
         sub_obj.delete()
         return common_response(code=200)
@@ -200,6 +232,10 @@ class SubjectView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Subject', 'PUT'):
+            return common_response(code=500, msg='您没有操作权限')
+
         data = request.data
         # 获取项目id
         sub_id = data.get('sub_id', 0)
@@ -210,18 +246,19 @@ class SubjectView(APIView):
         except Subject.DoesNotExist:
             return common_response(code=500, msg='ID不存在')
 
-        # 验证数据
-        subject_num = data.get('subject_num', '')
-        subject_name = data.get('subject_name', '')
-        subject_father = data.get('subject_father', None)
-        subject_type = data.get('subject_type', 0)
-        balance_type = data.get('balance_type', 0)
-        assists = data.get('assists', 0)
-        num = data.get('num', sub_obj.subjectofaccounts.num)
-        money = Decimal(data.get('money', sub_obj.subjectofaccounts.money)).quantize(Decimal('0.00'))
+        # 判断当前用户是否可操作
+        if not sub_obj.can_operation:
+            return common_response(code=500, msg='当前科目不可操作')
 
+        # 验证数据
+        subject_num = data.get('subject_num', sub_obj.subject_num)
+        subject_name = data.get('subject_name', sub_obj.subject_name)
+        subject_father = data.get('subject_father', None)
+        subject_type = data.get('subject_type', sub_obj.subject_type)
+        balance_type = data.get('balance_type', sub_obj.balance_type)
+        assists = data.get('assists', 0)
         # 数据入库
-        if not subject_num and not subject_name:
+        if not subject_num or not subject_name:
             return common_response(code=500, msg='缺少必要参数')
 
         # 获取父级对象
@@ -238,14 +275,11 @@ class SubjectView(APIView):
         sub_obj.parent = subject_father
         sub_obj.subject_type = subject_type
         sub_obj.balance_type = balance_type
-        sub_obj.subjectofaccounts.num = num
-        sub_obj.subjectofaccounts.money = money
 
         # 修改多对多关联表数据
         if assists:
             # 字符串转为列表
             ass_list = assists.strip(',').split(',')
-            # ass_obj = Assist.objects.filter(id__in=ass_list)
             sub_obj.assist_business.clear()
             sub_obj.assist_business.set(ass_list)
 
@@ -268,6 +302,10 @@ class SubjectView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Subject', 'GET'):
+            return common_response(code=500, msg='您没有操作权限')
+
         sub_id = request.GET.get('sub_id', '')
         data_type = request.GET.get('type', 'info')
 
@@ -281,6 +319,10 @@ class SubjectView(APIView):
             }
             ass_list.append(ass_info)
 
+        ''' 所有科目 '''
+        all_subjects = get_all_subject(obj)
+
+        # 获取详情
         if sub_id:
             try:
                 sub_obj = Subject.objects.get(id=sub_id)
@@ -295,35 +337,26 @@ class SubjectView(APIView):
                 'subject_type' : sub_obj.subject_type,
                 'balance_type' : sub_obj.balance_type,
                 'assists' : sub_obj.assist_list(),
-                'num' : sub_obj.subjectofaccounts.num if sub_obj.subjectofaccounts.num else 0,
-                'money' : sub_obj.subjectofaccounts.money if sub_obj.subjectofaccounts.money else 0,
+                'num' : sub_obj.subjectofaccounts.num if hasattr(sub_obj, 'subjectofaccounts') else 0,
+                'money' : sub_obj.subjectofaccounts.money if hasattr(sub_obj, 'subjectofaccounts') else 0,
                 'assists' : sub_obj.assist_list()
             }
             data = {
                 'sub_info': sub_info,
-                'ass_list': ass_list
+                'ass_list': ass_list,
+                'all_subjects': all_subjects
             }
             return common_response(data=data)
 
         if data_type == 'info':
-            return common_response(data=ass_list)
+            data = {
+                'all_subjects': all_subjects,
+                'ass_list': ass_list
+            }
+            return common_response(data=data)
 
         if data_type == 'list':
-            sub_all = Subject.objects.all()
-            sub_list = []
-            for sub_obj in sub_all:
-                sub_info = {
-                    'id': sub_obj.id,
-                    'subject_num': sub_obj.subject_num,
-                    'subject_name': sub_obj.subject_name,
-                    'parent': sub_obj.parent.subject_name if sub_obj.parent else '-',
-                    'subject_type': sub_obj.get_subject_type_display(),
-                    'balance_type': sub_obj.get_balance_type_display(),
-                    'num' : sub_obj.subjectofaccounts.num if sub_obj.subjectofaccounts.num else 0,
-                    'money' : sub_obj.subjectofaccounts.money if sub_obj.subjectofaccounts.money else 0,
-                }
-                sub_list.append(sub_info)
-            return common_response(data=sub_list)
+            return common_response(data=all_subjects)
 
 
 class AssisView(APIView):
@@ -336,9 +369,16 @@ class AssisView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Assis', 'POST'):
+            return common_response(code=500, msg='您没有操作权限')
+
         data = request.data
         ass_name = data['name'] if 'name' in data else ''
-        ass_obj = Assist.objects.create(assist_name=ass_name)
+        ass_obj = Assist.objects.create(
+            assist_name=ass_name,
+            village=obj.village
+        )
         ass_obj.save()
         return common_response(msg='True')
 
@@ -349,6 +389,10 @@ class AssisView(APIView):
             obj = User.objects.get(id=user_id)
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
+
+        # 验证用户权限
+        if not has_permission(obj, 'Assis', 'DELETE'):
+            return common_response(code=500, msg='您没有操作权限')
 
         data = request.data
         ass_id = data.get('ass_id', '')
@@ -367,6 +411,10 @@ class AssisView(APIView):
             obj = User.objects.get(id=user_id)
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
+
+        # 验证用户权限
+        if not has_permission(obj, 'Assis', 'PUT'):
+            return common_response(code=500, msg='您没有操作权限')
 
         data = request.data
         ass_id = data['ass_id'] if 'ass_id' in data else ''
@@ -389,6 +437,10 @@ class AssisView(APIView):
         except User.DoesNotExist:
             return common_response(code=500, msg='用户不存在!')
 
+        # 验证用户权限
+        if not has_permission(obj, 'Assis', 'GET'):
+            return common_response(code=500, msg='您没有操作权限')
+
         # 详情
         ass_id = request.GET.get('ass_id', '')
         if ass_id:
@@ -404,7 +456,7 @@ class AssisView(APIView):
                 return common_response(data=info)
 
         # 辅助核算列表
-        ass_obj = Assist.objects.all()
+        ass_obj = Assist.objects.all(village=obj.village)
         ass_list = []
         for a_obj in ass_obj:
             ass_info = {
@@ -447,34 +499,112 @@ class BalanceView(APIView):
 
         return common_response()
 
-# class SetSubjectMoneyView(APIView):
-#     def post(self, request):
-#         ''' 设置科目余额 '''
-#         # 验证用户信息
-#         is_log, user_id = is_logined(request)
-#         try:
-#             obj = User.objects.get(id=user_id)
-#         except User.DoesNotExist:
-#             return common_response(code=500, msg='用户不存在!')
 
-#         data = request.data
-#         soa_id = data.get('soa_id', 0)
-#         sub_data = data.get('sub_data', '')
-#         for child_data in sub_data:
-#             subject_id = data.get('subject_id', 0)
-#             num = data.get('num', 0)
-#             money = Decimal(data.get('money', 0)).quantize(Decimal('0.00'))
-#             if not soa_id or not subject_id:
-#                 return common_response(code=500, msg='缺少帐套id')
+class SetSubjectMoneyView(APIView):
+    def put(self, request):
+        # 验证用户信息
+        is_log, user_id = is_logined(request)
+        try:
+            obj = User.objects.get(id=user_id)
+        except User.DoesNotExist:
+            return common_response(code=500, msg='用户不存在!')
 
-#             soa_obj = SubjectOfAccounts.objects.create(
-#                 account_id=soa_id,
-#                 subject_id=subject_id,
-#                 num=num,
-#                 money=money
-#             )
-#             soa_obj.save()
-#         return common_response(msg='True')
+        # 验证用户权限
+        if not has_permission(obj, 'SetSubjectMoney', 'PUT'):
+            return common_response(code=500, msg='您没有操作权限')
+
+        # 验证帐套信息
+        data = request.data
+        soa_id = data.get('soa_id', 0)
+        if not soa_id:
+            return common_response(code=500, msg='缺少帐套id!')
+        try:
+            soa_obj = SetOfAccounts.objects.get(id=soa_id)
+        except SetOfAccounts.DoesNotExist:
+            return common_response(code=500, msg='帐套不存在!')
+
+        subjects = json.loads(data['subjects'])
+        for obj in subjects:
+            try:
+                subject_obj = Subject.objects.get(id=int(obj['id']))
+            except Subject.DoesNotExist:
+                continue
+
+            num = obj.get('num', 0)
+            money = Decimal(obj.get('money', 0)).quantize(Decimal('0.00'))
+            balance = num * money
+            if hasattr(subject_obj, 'subjectofaccounts'):
+                subject_obj.subjectofaccounts.num = num
+                subject_obj.subjectofaccounts.money = money
+                subject_obj.subjectofaccounts.balance = balance
+                subject_obj.subjectofaccounts.save()
+                continue
+            else:
+                soa_obj = SubjectOfAccounts.objects.create(
+                    account=soa_obj,
+                    subject=subject_obj,
+                    num=num,
+                    money=money,
+                    balance=balance
+                )
+                soa_obj.save()
+
+        return common_response(msg='True')
+
+    def get(self, request):
+        # 验证用户信息
+        is_log, user_id = is_logined(request)
+        try:
+            obj = User.objects.get(id=user_id)
+        except User.DoesNotExist:
+            return common_response(code=500, msg='用户不存在!')
+
+        # 验证用户权限
+        if not has_permission(obj, 'SetSubjectMoney', 'GET'):
+            return common_response(code=500, msg='您没有操作权限')
+
+        # 验证帐套信息
+        soa_id = request.GET.get('soa_id', 0)
+        if not soa_id:
+            return common_response(code=500, msg='缺少帐套id!')
+        try:
+            soa_obj = SetOfAccounts.objects.get(id=soa_id)
+        except SetOfAccounts.DoesNotExist:
+            return common_response(code=500, msg='帐套不存在!')
+
+        # 获取所有科目
+        all_subjects = get_all_subject(obj)
+        
+        return common_response(data=all_subjects)
+
+
+    # def post(self, request):
+    #     ''' 设置科目余额 '''
+    #     # 验证用户信息
+    #     is_log, user_id = is_logined(request)
+    #     try:
+    #         obj = User.objects.get(id=user_id)
+    #     except User.DoesNotExist:
+    #         return common_response(code=500, msg='用户不存在!')
+
+    #     data = request.data
+    #     soa_id = data.get('soa_id', 0)
+    #     sub_data = data.get('sub_data', '')
+    #     for child_data in sub_data:
+    #         subject_id = data.get('subject_id', 0)
+    #         num = data.get('num', 0)
+    #         money = Decimal(data.get('money', 0)).quantize(Decimal('0.00'))
+    #         if not soa_id or not subject_id:
+    #             return common_response(code=500, msg='缺少帐套id')
+
+    #         soa_obj = SubjectOfAccounts.objects.create(
+    #             account_id=soa_id,
+    #             subject_id=subject_id,
+    #             num=num,
+    #             money=money
+    #         )
+    #         soa_obj.save()
+    #     return common_response(msg='True')
 
 #     def put(self, request):
 #         # 验证用户信息
